+++ src/main/java/website/automate/manager/api/client/JobManagementRemoteService.java
@@ -1,14 +1,14 @@
-package website.automate.manager.pub.api.client;
+package website.automate.manager.api.client;
 
 import static java.util.Arrays.asList;
 
 import java.util.Collection;
 import java.util.List;
 
-import website.automate.manager.pub.api.client.model.Authentication;
-import website.automate.manager.pub.api.client.model.Job;
-import website.automate.manager.pub.api.client.model.Job.JobProfile;
-import website.automate.manager.pub.api.client.support.RestTemplate;
+import website.automate.manager.api.client.model.Authentication;
+import website.automate.manager.api.client.model.Job;
+import website.automate.manager.api.client.model.Job.JobProfile;
+import website.automate.manager.api.client.support.RestTemplate;
 
 public class JobManagementRemoteService {