+++ src/test/java/website/automate/manager/api/client/ScenarioRetrievalRemoteServiceIT.java
@@ -1,4 +1,4 @@
-package website.automate.manager.pub.api.client;
+package website.automate.manager.api.client;
 
 import static org.junit.Assert.*;
 
@@ -6,9 +6,10 @@ import java.util.List;
 
 import org.junit.Test;
 
-import website.automate.manager.pub.api.client.model.Authentication;
-import website.automate.manager.pub.api.client.model.Scenario;
-import website.automate.manager.pub.api.client.support.AuthenticationUtils;
+import website.automate.manager.api.client.ScenarioRetrievalRemoteService;
+import website.automate.manager.api.client.model.Authentication;
+import website.automate.manager.api.client.model.Scenario;
+import website.automate.manager.api.client.support.AuthenticationUtils;
 
 public class ScenarioRetrievalRemoteServiceIT extends MockServerIT {