+++ CHANGELOG.md
@@ -1,12 +1,15 @@
-# Automate Website Manager Public API Client Change Log
+# Automate Website Manager API Client Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed 
+- adjusted package names
+- shortened artifact name
 
 ## [0.0.2]
 ### Added
 - created initial implementation of the API
 
-[Unreleased]: https://github.com/automate-website/manager-public-api-client/compare/0.0.2...HEAD
-[0.0.2]: https://github.com/automate-website/manager-public-api-client/compare/0.0.0...0.0.2
+[Unreleased]: https://github.com/automate-website/manager-api-client/compare/0.0.2...HEAD
+[0.0.2]: https://github.com/automate-website/manager-api-client/compare/0.0.0...0.0.2